MPU cleanups, stack adjustments
-Cleaned up some of the constants and comments used for MPU setup.

-Changed stacks set up for the CPU modes to use the second ITCM nirror.
Maybe I should allow for programmers to provide the ITCM base they want
to use for this? This change was made so programmers can, later, set up
the MPU to make the first mirror be executable only, so any invalid
accesses to 0x0 and low memory leads to data aborts.